+++ Bitvavo.API/Models/Trade.cs
@@ -29,16 +29,16 @@ public class Trade
     public Side Side { get; set; }
 
     [JsonProperty("amount")]
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
 
     [JsonProperty("price")]
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 
     [JsonProperty("taker")]
     public bool Taker { get; set; }
 
     [JsonProperty("fee")]
-    public double Fee { get; set; }
+    public decimal Fee { get; set; }
 
     [JsonProperty("feeCurrency")]
     public string FeeCurrency { get; set; }
